requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
anthropic>=0.40.0
yt-dlp>=2024.1.0
openai-whisper>=20231117
Pillow>=10.0.0
pydantic>=2.0.0
